backend/requirements.txt
# Dependencias; con uv usa: uv sync (leyendo pyproject.toml)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
neo4j>=5.14.0
python-dotenv>=1.0.0
